summary: only show first line of Base.showerror()
Base.showerror() could print an arbitrarily long exception message. A
particularly long message can reduce the utility of the summary since
it detracts from the visual cues presented by the summary.

As a stopgap, only print the first line of the exception. This ensures
the summary is not spammy, while still showing the (presumptively) key
part of the exception.